Fix a few issues reported by PHPStan
Summary:
* scripts/arcanist.php:528    Parameter #2 $subject of function preg_match expects string, string|false given.
* scripts/arcanist.php:536    Parameter #1 $haystack of function strpos expects string, string|null given.
* src/channel/PhutilChannel.php:347    Method PhutilChannel::getReadSockets() has invalid return type stream.
* src/channel/PhutilChannel.php:359    Method PhutilChannel::getWriteSockets() has invalid return type stream.
* src/console/PhutilConsoleProgressBar.php:142    Parameter #2 $times of function str_repeat expects int, float given.
* src/channel/PhutilSocketChannel.php:50     Parameter #1 $stream of function stream_set_blocking expects resource, Socket given.
* src/init/lib/moduleutils.php:33     Parameter #1 $options of function debug_backtrace expects int, false given.
* src/parser/argument/PhutilArgumentParser.php:402    Method PhutilArgumentParser::parseWorkflowsFull() has invalid return type no.
* src/phage/agent/PhagePHPAgent.php:100    Parameter #1 $seconds of method ExecFuture::setTimeout() expects int, float given.
* src/runtime/ArcanistRuntime.php:261    Parameter #2 $subject of function preg_match expects string, string|false given.
* src/runtime/ArcanistRuntime.php:268    Parameter #1 $haystack of function strpos expects string, string|null given.
* src/workflow/ArcanistWorkflow.php:1877   Parameter #1 $string of function md5 expects string, string|false given.
* support/lib/extract-symbols.php:563    Parameter #1 $json of function json_decode expects string, string|false given.

No worries, future contributor, there are still many more issues left to fix.

Test Plan: Run static code analysis (PHPStan).

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26360

+++ src/phage/agent/PhagePHPAgent.php
@@ -97,7 +97,7 @@ final class PhagePHPAgent extends Phobject {
 
         $timeout = $spec['timeout'];
         if ($timeout) {
-          $future->setTimeout(ceil($timeout));
+          $future->setTimeout((int)ceil($timeout));
         }
 
         $future->setFutureKey($key);
